Added blank .vimrc.
Note that .vimrc is inside the .vim directory. This is because I want to keep
everything in one file for easier version control. To use this configuration
you still need to put a .vimrc or _vimrc (Windows) file in your home directory
with a 'source ~/.vim/.vimrc' or 'source ~/vimfiles/.vimrc' (Windows) line in
it.